+++ lib/models/invoice_item.dart
@@ -17,7 +17,7 @@ class InvoiceItem{
   final Stock stock;
   final Product product;
   final GRN grn;
-   int quantity;
+   double quantity;
   final double unitPrice;
   final double discount;
 
@@ -115,7 +115,7 @@ class InvoiceItem{
         id:int.parse(row.colByName(GRN.COLNAME_ID) as String),
         barcode: row.colByName(GRN.COLNAME_BARCODE)as String,
         grnDate: DateTime.parse(row.colByName(GRN.COLNAME_GRN_DATE)as String),
-        quantity: int.parse(row.colByName(GRN.COLNAME_QUANTITY)as String),
+        quantity: double.parse(row.colByName(GRN.COLNAME_QUANTITY)as String),
         description: row.colByName(GRN.COLNAME_DESCRIPTION)as String,
         mnfDate: DateTime.parse(row.colByName(GRN.COLNAME_MNF_DATE)as String),
         expDate: DateTime.parse(row.colByName(GRN.COLNAME_EXP_DATE)as String),
@@ -155,7 +155,7 @@ class InvoiceItem{
       Stock stock = Stock(
           id: int.parse(row.colByName(Stock.COLNAME_ID)??"0"),
           barcode: row.colByName(Stock.COLNAME_BARCODE) as String,
-          availbleQty: int.parse(row.colByName(Stock.COLNAME_AVAILBLE_QTY)?? "0"),
+          availbleQty: double.parse(row.colByName(Stock.COLNAME_AVAILBLE_QTY)?? "0"),
           defaultDiscount: double.parse(row.colByName(Stock.COLNAME_DEFAULT_DISCOUNT)?? "0"),
           retailPrice: double.parse(row.colByName(Stock.COLNAME_RETAIL_PRICE)?? "0"),
           wholesalePrice: double.parse(row.colByName(Stock.COLNAME_WHOLESALE_PRICE)?? "0"),
@@ -168,7 +168,7 @@ class InvoiceItem{
 
       InvoiceItem invoiceItem = InvoiceItem(
         invoiceID: int.parse(row.colByName(COLNAME_INVOICE_ID) as String),
-        quantity: int.parse(row.colByName(COLNAME_QUANTITY) as String),
+        quantity: double.parse(row.colByName(COLNAME_QUANTITY) as String),
         unitPrice: double.parse(row.colByName(COLNAME_UNIT_PRICE) as String),
         product: product,
         stock: stock,
